+++ src/Fuzz/Command/AbstractHashReadCommand.php
@@ -23,7 +23,7 @@ abstract class AbstractHashReadCommand extends RedisCommand
 
     public function createOperation(FuzzContext $context): RedisOperation
     {
-        $key = $context->randomKey($this->type());
+        $key = $context->randomKey($this->type(), $this->flags());
 
         return new RedisOperation(
             $this->name(),