+++ src/Fuzz/Command/AbstractSingleKeyCommand.php
@@ -11,7 +11,7 @@ abstract class AbstractSingleKeyCommand extends RedisCommand
 {
     public function createOperation(FuzzContext $context): RedisOperation
     {
-        $key = $context->randomKey($this->type());
+        $key = $context->randomKey($this->type(), $this->flags());
 
         return new RedisOperation(
             $this->name(),