+++ src/Fuzz/Command/AbstractSingleKeyValueCommand.php
@@ -17,7 +17,7 @@ abstract class AbstractSingleKeyValueCommand extends RedisCommand
 
     public function createOperation(FuzzContext $context): RedisOperation
     {
-        $key = $context->randomKey($this->type());
+        $key = $context->randomKey($this->type(), $this->flags());
 
         return new RedisOperation(
             $this->name(),