+++ src/Fuzz/Command/AbstractSortedSetWriteCommand.php
@@ -23,7 +23,7 @@ abstract class AbstractSortedSetWriteCommand extends RedisCommand
 
     public function createOperation(FuzzContext $context): RedisOperation
     {
-        $key = $context->randomKey($this->type());
+        $key = $context->randomKey($this->type(), $this->flags());
         $arguments = [$key];
 
         foreach ($context->randomSortedSetEntries() as $entry) {