+++ src/Fuzz/Command/Definitions/ZRangeCommand.php
@@ -30,7 +30,7 @@ final class ZRangeCommand extends AbstractSingleKeyCommand
 
     public function createOperation(FuzzContext $context): RedisOperation
     {
-        $key = $context->randomKey($this->type());
+        $key = $context->randomKey($this->type(), $this->flags());
 
         return new RedisOperation(
             $this->name(),